requirements.txt
flet-cli==0.25.1
  - cookiecutter [required: >=2.6.0,<3.0.0, installed: 2.6.0]
    - arrow [required: Any, installed: 1.3.0]
      - python-dateutil [required: >=2.7.0, installed: 2.9.0.post0]
        - six [required: >=1.5, installed: 1.16.0]
      - types-python-dateutil [required: >=2.8.10, installed: 2.9.0.20241003]
    - binaryornot [required: >=0.4.4, installed: 0.4.4]
      - chardet [required: >=3.0.2, installed: 5.2.0]
    - click [required: >=7.0,<9.0.0, installed: 8.1.7]
      - colorama [required: Any, installed: 0.4.6]
    - Jinja2 [required: >=2.7,<4.0.0, installed: 3.1.4]
      - MarkupSafe [required: >=2.0, installed: 3.0.2]
    - python-slugify [required: >=4.0.0, installed: 8.0.4]
      - text-unidecode [required: >=1.3, installed: 1.3]
    - PyYAML [required: >=5.3.1, installed: 6.0.2]
    - requests [required: >=2.23.0, installed: 2.32.3]
      - certifi [required: >=2017.4.17, installed: 2024.8.30]
      - charset-normalizer [required: >=2,<4, installed: 3.4.0]
      - idna [required: >=2.5,<4, installed: 3.10]
      - urllib3 [required: >=1.21.1,<3, installed: 2.2.3]
    - rich [required: Any, installed: 13.9.4]
      - markdown-it-py [required: >=2.2.0, installed: 3.0.0]
        - mdurl [required: ~=0.1, installed: 0.1.2]
      - Pygments [required: >=2.13.0,<3.0.0, installed: 2.18.0]
      - typing_extensions [required: >=4.0.0,<5.0, installed: 4.12.2]
  - flet [required: ==0.25.1, installed: 0.25.1]
    - httpx [required: Any, installed: 0.28.0]
      - anyio [required: Any, installed: 4.6.2.post1]
        - exceptiongroup [required: >=1.0.2, installed: 1.2.2]
        - idna [required: >=2.8, installed: 3.10]
        - sniffio [required: >=1.1, installed: 1.3.1]
        - typing_extensions [required: >=4.1, installed: 4.12.2]
      - certifi [required: Any, installed: 2024.8.30]
      - httpcore [required: ==1.*, installed: 1.0.7]
        - certifi [required: Any, installed: 2024.8.30]
        - h11 [required: >=0.13,<0.15, installed: 0.14.0]
      - idna [required: Any, installed: 3.10]
    - oauthlib [required: >=3.2.2,<4.0.0, installed: 3.2.2]
    - repath [required: >=0.9.0,<0.10.0, installed: 0.9.0]
      - six [required: >=1.9.0, installed: 1.16.0]
  - packaging [required: Any, installed: 24.2]
  - qrcode [required: >=7.4.2,<8.0.0, installed: 7.4.2]
    - colorama [required: Any, installed: 0.4.6]
    - pypng [required: Any, installed: 0.20220715.0]
    - typing_extensions [required: Any, installed: 4.12.2]
  - toml [required: >=0.10.2,<0.11.0, installed: 0.10.2]
  - watchdog [required: >=4.0.0,<5.0.0, installed: 4.0.2]
flet-desktop==0.25.1
  - flet [required: ==0.25.1, installed: 0.25.1]
    - httpx [required: Any, installed: 0.28.0]
      - anyio [required: Any, installed: 4.6.2.post1]
        - exceptiongroup [required: >=1.0.2, installed: 1.2.2]
        - idna [required: >=2.8, installed: 3.10]
        - sniffio [required: >=1.1, installed: 1.3.1]
        - typing_extensions [required: >=4.1, installed: 4.12.2]
      - certifi [required: Any, installed: 2024.8.30]
      - httpcore [required: ==1.*, installed: 1.0.7]
        - certifi [required: Any, installed: 2024.8.30]
        - h11 [required: >=0.13,<0.15, installed: 0.14.0]
      - idna [required: Any, installed: 3.10]
    - oauthlib [required: >=3.2.2,<4.0.0, installed: 3.2.2]
    - repath [required: >=0.9.0,<0.10.0, installed: 0.9.0]
      - six [required: >=1.9.0, installed: 1.16.0]
loguru==0.7.3
  - colorama [required: >=0.3.4, installed: 0.4.6]
  - win32_setctime [required: >=1.0.0, installed: 1.2.0]
music21==9.3.0
  - chardet [required: Any, installed: 5.2.0]
  - joblib [required: Any, installed: 1.4.2]
  - jsonpickle [required: Any, installed: 4.0.0]
  - matplotlib [required: Any, installed: 3.9.3]
    - contourpy [required: >=1.0.1, installed: 1.3.1]
      - numpy [required: >=1.23, installed: 2.1.3]
    - cycler [required: >=0.10, installed: 0.12.1]
    - fonttools [required: >=4.22.0, installed: 4.55.2]
    - kiwisolver [required: >=1.3.1, installed: 1.4.7]
    - numpy [required: >=1.23, installed: 2.1.3]
    - packaging [required: >=20.0, installed: 24.2]
    - pillow [required: >=8, installed: 11.0.0]
    - pyparsing [required: >=2.3.1, installed: 3.2.0]
    - python-dateutil [required: >=2.7, installed: 2.9.0.post0]
      - six [required: >=1.5, installed: 1.16.0]
  - more-itertools [required: Any, installed: 10.5.0]
  - numpy [required: Any, installed: 2.1.3]
  - requests [required: Any, installed: 2.32.3]
    - certifi [required: >=2017.4.17, installed: 2024.8.30]
    - charset-normalizer [required: >=2,<4, installed: 3.4.0]
    - idna [required: >=2.5,<4, installed: 3.10]
    - urllib3 [required: >=1.21.1,<3, installed: 2.2.3]
  - webcolors [required: >=1.5, installed: 24.11.1]